added friends notes and follow buttons

+++ classes/UploadedProfileImageFile.php
@@ -20,7 +20,7 @@ class UploadedProfileImageFile extends UploadedImageFile {
 
         parent::__construct($uploadedFile); //super constructor 
         $this->filename = $_SESSION["user"] . "-" . filemtime($this->tempFilePath); //<username>-<timestamp> as filename, where timestamp is unix upload time
-        $this->mysql = MySQL::getinstance();
+        $this->mysql = MySQL::getInstance();
 
     }
 
@@ -39,9 +39,15 @@ class UploadedProfileImageFile extends UploadedImageFile {
             
             if ($deleteExisting) { //if existing file should be removed, get the existing path, delete it after successful persistance
                         
-                $oldFilePath = $this->mysql->request($this->mysql->readBasicProfileQuery, [":user" => $_SESSION["user"]])[0]["profilePictureURL"];
+                $oldFilePath = $this->mysql->request($this->mysql->readBasicProfileQuery, [":user" => $_SESSION["user"]])[0]["profilePictureURL"]; //abs path to existing file
+                $oldFileName = basename($oldFilePath); //filename including ext
+                $noDeleteFiles = ["avatar0.png", "avatar1.png", "avatar2.png", "avatar3.png", "avatar4.png", "avatar5.png", "avatar6.png", "avatar7.png", "avatar8.png"]; //files that should not be deleted
                 
                 $this->persist();
+
+                if (in_array($oldFileName, $noDeleteFiles)) {
+                    return true;
+                }
 
                 if (!unlink($oldFilePath)) {
                     error_log("Failed to delete a profile photo: " . $oldFilePath);
